requirements.txt
Django==2.0.4
django-configurations==2.0
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.7.1
psycopg2==2.7.4
requests==2.18.4
urllib3==1.22
base32-crockford==0.3.0
django-rq==1.0.1
rq==0.10.0
-e git+https://github.com/dankolbman/hvac#egg=hvac
drf-yasg==1.6.1
boto==2.48.0
PyJWT==1.6.4
cryptography==2.3.0
django-cors-headers==2.2.0
django-fsm==2.6.0
semantic-version==2.6.0
drf-nested-routers==0.90.2
